storeBackend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers
psycopg2-binary==2.9.9
pillow==10.4.0